requirements.txt
aioredis==1.3.1
ariadne==0.12.0
async-timeout==3.0.1
click==7.1.2
graphql-core==3.0.5
h11==0.10.0
hiredis==1.1.0
passlib==1.7.2
pydantic==1.6.1
PyJWT==1.7.1
SQLAlchemy==1.3.19
starlette==0.13.8
typing-extensions==3.7.4.3
uvicorn==0.12.1
